Refactor player queues: PlayerQueueData record + concern subcontrollers
Consolidate the per-queue runtime state that was spread across ~8 parallel
queue_id-keyed dicts/sets on the controller into one server-side
PlayerQueueData record (state.py), which also owns the cache (de)serialization.
Fold the dynamic-source items onto it (persisted directly) and drop the lossy
_restore_source_items rebuild; fix a stale _flow_buffer_completed entry on
player removal.

Slice the controller into concern-focused subcontrollers it constructs and
delegates to: MediaResolver (media->tracks), QueueLoader (enqueue + fill),
PlaybackTracker (player->queue reconciliation + play-counting) and StreamFeeder
(next-item / buffer feed). Rename the existing helper objects to drop the
Helper suffix (Autoplay/SmartShuffle/ManagedPool). The controller keeps the
public API commands and inter-controller event hooks and dispatches the logic.

No behaviour change; tests updated/added and the package README refreshed.
