+++ music_assistant/controllers/player_queues/managed_pool.py
@@ -94,7 +94,7 @@ class _MaterializedSource:
     fully_paged: bool
 
 
-class ManagedPoolHelper:
+class ManagedPool:
     """Build and top up a queue's bounded managed pool from its dynamic sources."""
 
     def __init__(self, queues: PlayerQueuesController) -> None:
@@ -116,7 +116,7 @@ class ManagedPoolHelper:
         :param queue_id: The queue to fill.
         :param is_initial: True when seeding a fresh pool, False when topping up an existing one.
         """
-        queue = self.queues._queues[queue_id]
+        queue = self.queues._queue_data[queue_id].queue
         windows = self.queues._smart_shuffle._windows(queue_id)
         snapshot = await self.mass.music.recency.snapshot(windows, userid=queue.userid)
         # publish a best-effort recency filter so dynamic-playlist generation can pre-skip recently
@@ -129,7 +129,7 @@ class ManagedPoolHelper:
             return []
         # dedupe only against the active (current + unplayed) tail; played history is left out so
         # recency, not permanent exclusion, decides when a track may return
-        items = self.queues._queue_items[queue_id]
+        items = self.queues._queue_data[queue_id].items
         start = queue.current_index if queue.current_index is not None else 0
         pool_keys = {
             item.media_item for item in items[start:] if isinstance(item.media_item, Track)
@@ -181,7 +181,7 @@ class ManagedPoolHelper:
         # multiplicity = how often a source was added (adding a source more than once weights it up)
         counts: dict[str, int] = {}
         items: dict[str, MediaItemType] = {}
-        for item in self.queues._source_items.get(queue_id, []):
+        for item in self.queues._queue_data[queue_id].source_items:
             if isinstance(item, Playlist) and item.is_dynamic and not include_dynamic:
                 continue
             if not (uri := _uri(item)):
@@ -222,14 +222,16 @@ class ManagedPoolHelper:
         if not isinstance(media_item, Playlist):
             return []
         with suppress(MusicAssistantError):
-            tracks = await self.queues.get_playlist_tracks(media_item, start_item=None)
+            tracks = await self.queues._media_resolver.get_playlist_tracks(
+                media_item, start_item=None
+            )
             return [track for track in tracks if isinstance(track, Track) and track.available]
         return []
 
     async def _fetch_tracks(self, media_item: MediaItemType) -> list[Track]:
         """Fetch a TRACKS source's own (playable) tracks, honoring the user's selection prefs."""
         with suppress(MusicAssistantError):
-            tracks = await self.queues.get_tracks_for_playback(media_item)
+            tracks = await self.queues._media_resolver.get_tracks_for_playback(media_item)
             return [track for track in tracks if isinstance(track, Track) and track.available]
         return []
 
@@ -306,12 +308,14 @@ class ManagedPoolHelper:
 
     def _retire_sources(self, queue_id: str, uris: list[str]) -> None:
         """Drop fully-played finite sources from the queue so only live sources remain."""
-        queue = self.queues._queues.get(queue_id)
+        queue = self.queues.get(queue_id)
         if queue is None:
             return
         retire = set(uris)
         remaining = [
-            item for item in self.queues._source_items.get(queue_id, []) if _uri(item) not in retire
+            item
+            for item in self.queues._queue_data[queue_id].source_items
+            if _uri(item) not in retire
         ]
         self.queues._store_sources(queue, remaining)
         self.logger.debug(
@@ -321,7 +325,7 @@ class ManagedPoolHelper:
 
     def _unplayed(self, queue: PlayerQueue) -> int:
         """Return how many not-yet-played items remain in the queue."""
-        items = self.queues._queue_items.get(queue.queue_id, [])
+        items = data.items if (data := self.queues._queue_data.get(queue.queue_id)) else []
         if queue.current_index is None:
             return len(items)
         return max(len(items) - (queue.current_index + 1), 0)
