+++ music_assistant/controllers/streams/audio.py
@@ -1614,7 +1614,9 @@ class StreamsAudio:
                     >= streamdetails.duration - 60
                 ):
                     next_buffer_triggered = True
-                    self.mass.player_queues._prepare_next_audio_buffer(queue_item.queue_id)
+                    self.mass.player_queues._stream_feeder._prepare_next_audio_buffer(
+                        queue_item.queue_id
+                    )
                 yield chunk
                 del chunk
             # if we received no audio and the buffer has a producer error,
